+++ components/atoms/PullRequestOverviewChart/pull-request-overview-chart.tsx
@@ -7,7 +7,7 @@ interface PullRequestOverviewChartBarProps {
 
 const PullRequestOverviewChartBar: React.FC<PullRequestOverviewChartBarProps> = ({ percent, type }) => {
   return (
-    <div 
+    <div
       className={`
         ${ type === "open" ? "bg-light-grass-9" : type === "merged" ? "bg-purple-600" : type === "closed" ? "bg-light-red-9" : "bg-light-slate-9"}
         transition-all duration-500 ease-in-out rounded-full`}
@@ -24,6 +24,7 @@ interface PullRequestOverviewChartProps {
   totalPullRequests?: number;
 }
 
+
 const PullRequestOverviewChart: React.FC<PullRequestOverviewChartProps> = ({ className, open, merged, closed, draft, totalPullRequests }) => {  
   const getPercentage = (PullRequestType: number | undefined) => {
     return (((!!PullRequestType ? PullRequestType : 0) * 100) / (!!totalPullRequests ? totalPullRequests : 0));
@@ -32,18 +33,26 @@ const PullRequestOverviewChart: React.FC<PullRequestOverviewChartProps> = ({ cla
   return (
     <div
       className="w-full h-1.5 flex gap-0.5 bg-light-slate-2 rounded-full overflow-hidden">
-      
+
       {/* Open */}
-      <PullRequestOverviewChartBar percent={getPercentage(open)} type="open" />
+      { open && open > 0 && 
+        <PullRequestOverviewChartBar percent={getPercentage(open)} type="open" />
+      }
 
       {/* Merged */}
-      <PullRequestOverviewChartBar percent={getPercentage(merged)} type="merged" />
+      { merged && merged > 0 && 
+        <PullRequestOverviewChartBar percent={getPercentage(merged)} type="merged" />
+      }
 
       {/* Closed */}
-      <PullRequestOverviewChartBar percent={getPercentage(closed)} type="closed" />
-      
+      { closed && closed > 0 && 
+        <PullRequestOverviewChartBar percent={getPercentage(closed)} type="closed" />
+      }
+
       {/* Draft */}
-      <PullRequestOverviewChartBar percent={getPercentage(draft)} type="draft" />
+      { draft && draft > 0 && 
+        <PullRequestOverviewChartBar percent={getPercentage(draft)} type="draft" />
+      }
     </div>
   );
 };