+++ components/organisms/Dashboard/dashboard.tsx
@@ -4,26 +4,88 @@ import DashboardScatterChart from "components/molecules/DashboardScatterChart/da
 import HighlightCard from "components/molecules/HighlightCard/highlight-card";
 import humanizeNumber from "lib/utils/humanizeNumber";
 import { useEffect, useState } from "react";
+import { useContributionsList } from "lib/hooks/useContributionsList";
+import { calcDaysFromToday } from "lib/utils/date-utils";
+import { useMediaQuery } from "lib/hooks/useMediaQuery";
 
 export const Dashboard = (): JSX.Element => {
-  const { meta, isError } = useRepositoriesList();
+  // This is mock data for the dashboard. Not intended to be the final implementation.
+  const { data: contributorData, isError: contributorError } = useContributionsList("35");
 
+  const { meta: repoMetaData, isError: repoError } = useRepositoriesList();
   const [itemCountText, setItemCountText] = useState("Loading...");
+  const isNotMobile = useMediaQuery("(min-width: 768px)");
+
+  const conAvatarObject: { [key: string]: {[key: string]: string} } = {};
+
+  const fakeDataSet = [
+    33,
+    400,
+    12,
+    5049,
+    0,
+    840,
+    3603,
+    400,
+    220,
+    5,
+    1284,
+    7000,
+    1060,
+    64,
+    8099,
+    6400,
+    1234,
+    123,
+    802,
+    6000,
+    100,
+    1206,
+    2084,
+    786,
+    876,
+    954,
+    305,
+    1087,
+    2803,
+    55,
+    2,
+    103,
+    2,
+    902,
+    500,
+    702
+  ];
+
+  const scatterChartData = contributorError ? [] :
+    //eslint-disable-next-line
+    contributorData.map(({ last_commit_time, files_modified, host_login }, index) => {
+      const timeOverTouched: (string | number)[] = [
+        calcDaysFromToday(new Date(parseInt(last_commit_time))),
+        //eslint-disable-next-line
+        files_modified !== null ? files_modified : fakeDataSet[index]
+      ];
+
+      //eslint-disable-next-line
+      conAvatarObject[`${timeOverTouched[0]}${timeOverTouched[1]}`] = { login: host_login, image: `https://www.github.com/${host_login}.png?size=60` };
+
+      return timeOverTouched;
+    });
 
   const scatterOptions = {
     grid: {
       left: 40,
       top: 10,
       right: 40,
-      bottom: 20
+      bottom: 40
     },
     xAxis: {
       boundaryGap: false,
       scale: true,
-      minInterval: 7,
+      minInterval: isNotMobile ? 7 : 2,
       maxInterval: 7,
       min: 0,
-      max: 35,
+      max: isNotMobile ? 35 : 7,
       axisLabel: {
         formatter: (value: number, index: number) =>
           value === 0 ? "Today" : value === 35 ? "35+ days ago" : `${value} days ago`
@@ -39,6 +101,12 @@ export const Dashboard = (): JSX.Element => {
       max: 10000,
       splitNumber: 6,
       boundaryGap: false,
+      axisLine: {
+        show: false
+      },
+      axisTick: {
+        show: false
+      },
       axisLabel: {
         showMinLabel: true,
         formatter: (value: number) => value >= 1000 ? humanizeNumber(value,null) : value
@@ -49,43 +117,28 @@ export const Dashboard = (): JSX.Element => {
         }
       }
     },
+    tooltip: {
+      trigger: "item",
+      formatter: (args: any) => `${conAvatarObject[`${args.value[0]}${args.value[1]}`].login}`
+    },
     series: [
       {
-        symbolSize: 20,
-        data: [
-          [10.0, 8.04],
-          [8.07, 6.95],
-          [13.0, 7.58],
-          [9.05, 8.81],
-          [11.0, 8.33],
-          [14.0, 7.66],
-          [13.4, 6.81],
-          [10.0, 6.33],
-          [14.0, 8.96],
-          [12.5, 6.82],
-          [9.15, 7.2],
-          [11.5, 7.2],
-          [3.03, 4.23],
-          [12.2, 7.83],
-          [2.02, 4.47],
-          [1.05, 3.33],
-          [4.05, 4.96],
-          [6.03, 7.24],
-          [12.0, 6.26],
-          [12.0, 8.84],
-          [7.08, 5.82],
-          [5.02, 5.68]
-        ],
-        type: "scatter"
+        symbolSize: 40,
+        symbol: (value: number[]) => `image://${conAvatarObject[`${value[0]}${value[1]}`].image}`,
+        symbolOffset: [0, "-50%"],
+        data: scatterChartData,
+        type: "scatter",
+        itemStyle: {
+          opacity: 1
+        }
       }
     ]
   };
 
   useEffect(() => {
-
-    if(meta) setItemCountText(`of ${humanizeNumber(meta.itemCount, "comma")}`);
-    if(isError) setItemCountText("of unknown...");
-  }, [ isError, meta ]);
+    if (repoMetaData) setItemCountText(`of ${humanizeNumber(repoMetaData.itemCount, "comma")}`);
+    if (repoError) setItemCountText("of unknown...");
+  }, [repoError, repoMetaData]);
 
   return (
     <div className="flex flex-col w-full gap-4">