+++ interfaces/global-state-types.ts
@@ -2,4 +2,5 @@ import { User } from "@supabase/gotrue-js/src/lib/types";
 
 export interface GlobalStateInterface {
   user?: User | null;
+  onboarded?: boolean;
 }